# Eliminate copy in ForceCW and ForceCCW